+++ Assets/DEV/YJE/MissionBox.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c3487cb9a4267fa4481fe4075562cbc9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
