+++ Assets/DEV/YJE/MissionController.cs
@@ -9,22 +9,6 @@ public class MissionController : MonoBehaviour
     public bool Is2Clear = false;
     public bool IsEndingClear = false;
 
-    public bool IsUIOpen;
-
-    public void MissionBoxOpen()
-    {
-        Debug.Log("미션상자 열기");
-        IsUIOpen = true;
-        // TODO : UI 상호작용 창 닫혀있는지 확인하는 bool변수 = ture; - return값
-    }
-    public void MissionBoxClose()
-    {
-        Debug.Log("미션상자 닫기");
-        IsUIOpen = false;
-        // TODO : UI 상호작용 창 닫혀있는지 확인하는 bool변수 = false; - return값
-    }
-
-
     // TODO: UI와의 버튼으로 구현하게 될 예정
     /// <summary>
     /// Mission1의 클리어 여부를 확인하는 함수
