--- Assets/DEV/YJE/Prefabs/MissionBox2.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: a28549382ecfce64289b54888614d623
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
