+++ Assets/DEV/YJE/Prefabs/MissionController.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 64929e2a5c1f00e4e97f6b2e45f2bcfb
+guid: 5cf6998d868dd90408d14545c325bd41
 PrefabImporter:
   externalObjects: {}
   userData:
